+++ test/unit/specs/components/PageRender.spec.js
@@ -1,5 +1,5 @@
 import { mount } from 'vue-test-utils'
-import Component from '@/components/HelloWorld'
+import Component from '@/components/PageRender'
 
 describe(`${Component.name}.vue`, () => {
   const createCmp = propsData => mount(Component, { propsData })
@@ -8,6 +8,6 @@ describe(`${Component.name}.vue`, () => {
   it('show the correct message', () => {
     const cmp = createCmp(propsMock)
     const classes = cmp.classes()
-    expect(classes).toContain(`HelloWorld`)
+    expect(classes).toContain(`PageRender`)
   })
 })