+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Updated vuejs webpack template from 1.2.7 to 1.2.8
 - Removed CSS from base app component
 - Removed demo components from `HelloWorld`and updated tests
+- Rename HelloWorld to PageRender
 ### Deprecated
 ### Removed
 ### Fixed